v0.8.2 fix like key ampping to include missing keys
LikeKeys maps a set of options to a format rokaki can use
when building filters. Some of the data structure wasn't handled so keys
we being missed.

This fix makes use of 'undigging' (or deep assigning hashes) in
combination with a 'key path' array, to ensure all paths are covered.